Split tests in three files
